Simplification: Mount filesystems earlier in init, and don't remount on LoadExec
This is needed for an upcoming PR.
